+++ bashlib64-core.bash
@@ -4,7 +4,7 @@
 #
 # Author: serdigital64 (https://github.com/serdigital64)
 # Repository: https://github.com/serdigital64/bashlib64
-# Version: 10.0.0
+# Version: 10.1.0
 #
 # Copyright 2022 SerDigital64@gmail.com
 #
@@ -592,7 +592,7 @@ export _BL64_MSG_TXT_BATCH_FINISH_ERROR='finished with errors'
 #######################################
 # BashLib64 / Module / Globals / OS / Identify OS attributes and provide command aliases
 #
-# Version: 1.17.0
+# Version: 1.18.0
 #######################################
 
 export BL64_OS_MODULE="$BL64_VAR_OFF"
@@ -641,7 +641,8 @@ export _BL64_OS_TXT_CHECK_OS_MATRIX='Please check that the OS is listed in the c
 export _BL64_OS_TXT_FAILED_TO_NORMALIZE_OS='Unable to normalize OS name and version from /etc/os-release'
 export _BL64_OS_TXT_INVALID_OS_PATTERN='invalid OS pattern'
 export _BL64_OS_TXT_OS_NOT_SUPPORTED='BashLib64 is not supported in the current OS'
-
+export _BL64_OS_TXT_OS_VERSION_NOT_SUPPORTED='the current OS version is not compatible with the task'
+export _BL64_OS_TXT_OS_MATRIX='Supported OS Versions'
 
 #######################################
 # BashLib64 / Module / Globals / Manage role based access service
@@ -1019,7 +1020,7 @@ function bl64_check_setup() {
 #######################################
 # BashLib64 / Module / Functions / Check for conditions and report status
 #
-# Version: 3.0.0
+# Version: 3.0.2
 #######################################
 
 #######################################
@@ -1247,6 +1248,8 @@ function bl64_check_export() {
   bl64_dbg_lib_show_function "$@"
   local export_name="${1:-}"
   local description="${2:-export: $export_name}"
+
+  bl64_check_parameter 'export_name' || return $?
 
   if [[ ! -v "$export_name" ]]; then
     bl64_msg_show_error "${_BL64_CHECK_TXT_EXPORT_SET} (${description} ${BL64_MSG_COSMETIC_PIPE} ${_BL64_CHECK_TXT_FUNCTION}: ${FUNCNAME[1]:-NONE}@${BASH_LINENO[1]:-NONE})"
@@ -1529,10 +1532,10 @@ function bl64_check_alert_module_setup() {
 }
 
 #######################################
-# Check that parameters are passed
+# Check that at least 1 parameter is passed when using dynamic arguments
 #
 # Arguments:
-#   $1: total number of parameters from the calling function ($#)
+#   $1: must be $# to capture number of parameters from the calling function
 # Outputs:
 #   STDOUT: none
 #   STDERR: message
@@ -1541,7 +1544,9 @@ function bl64_check_alert_module_setup() {
 #######################################
 function bl64_check_parameters_none() {
   bl64_dbg_lib_show_function "$@"
-  local count="${1:-0}"
+  local count="$1"
+
+  bl64_check_parameter 'count' || return $?
 
   if [[ "$count" == '0' ]]; then
     bl64_msg_show_error "${_BL64_CHECK_TXT_NOARGS} (${_BL64_CHECK_TXT_FUNCTION}: ${FUNCNAME[1]:-NONE}@${BASH_LINENO[1]:-NONE})"
@@ -2357,7 +2362,7 @@ function _bl64_fs_set_alias() {
 #######################################
 # BashLib64 / Module / Functions / Manage local filesystem
 #
-# Version: 3.1.0
+# Version: 3.2.0
 #######################################
 
 #######################################
@@ -2585,9 +2590,9 @@ function bl64_fs_merge_dir() {
 #######################################
 function bl64_fs_run_chown() {
   bl64_dbg_lib_show_function "$@"
-  bl64_check_parameters_none "$#" || return $?
   local verbose=''
 
+  bl64_check_parameters_none "$#" || return $?
   bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_CHOWN_VERBOSE"
 
   bl64_dbg_lib_trace_start
@@ -2609,9 +2614,9 @@ function bl64_fs_run_chown() {
 #######################################
 function bl64_fs_run_chmod() {
   bl64_dbg_lib_show_function "$@"
-  bl64_check_parameters_none "$#" || return $?
   local verbose=''
 
+  bl64_check_parameters_none "$#" || return $?
   bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_CHMOD_VERBOSE"
 
   bl64_dbg_lib_trace_start
@@ -2722,9 +2727,9 @@ function bl64_fs_ln_symbolic() {
 #######################################
 function bl64_fs_run_mkdir() {
   bl64_dbg_lib_show_function "$@"
-  bl64_check_parameters_none "$#" || return $?
   local verbose=''
 
+  bl64_check_parameters_none "$#" || return $?
   bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_MKDIR_VERBOSE"
 
   bl64_dbg_lib_trace_start
@@ -2763,9 +2768,9 @@ function bl64_fs_mkdir_full() {
 #######################################
 function bl64_fs_run_mv() {
   bl64_dbg_lib_show_function "$@"
-  bl64_check_parameters_none "$#" || return $?
   local verbose=''
 
+  bl64_check_parameters_none "$#" || return $?
   bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_MV_VERBOSE"
 
   bl64_dbg_lib_trace_start
@@ -2787,12 +2792,10 @@ function bl64_fs_run_mv() {
 #######################################
 function bl64_fs_rm_file() {
   bl64_dbg_lib_show_function "$@"
-  local verbose=''
 
-  bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_RM_VERBOSE"
+  bl64_check_parameters_none "$#" || return $?
 
-  # shellcheck disable=SC2086
-  "$BL64_FS_CMD_RM" $verbose "$BL64_FS_SET_RM_FORCE" "$@"
+  bl64_fs_run_rm "$BL64_FS_SET_RM_FORCE" "$@"
 }
 
 #######################################
@@ -2808,12 +2811,10 @@ function bl64_fs_rm_file() {
 #######################################
 function bl64_fs_rm_full() {
   bl64_dbg_lib_show_function "$@"
-  local verbose=''
 
-  bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_RM_VERBOSE"
+  bl64_check_parameters_none "$#" || return $?
 
-  # shellcheck disable=SC2086
-  "$BL64_FS_CMD_RM" $verbose "$BL64_FS_SET_RM_FORCE" "$BL64_FS_SET_RM_RECURSIVE" "$@"
+  bl64_fs_run_rm "$BL64_FS_SET_RM_FORCE" "$BL64_FS_SET_RM_RECURSIVE" "$@"
 }
 
 #######################################
@@ -3179,9 +3180,9 @@ function bl64_fs_fix_permissions() {
 #######################################
 function bl64_fs_run_cp() {
   bl64_dbg_lib_show_function "$@"
-  bl64_check_parameters_none "$#" || return $?
   local verbose=''
 
+  bl64_check_parameters_none "$#" || return $?
   bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_CP_VERBOSE"
 
   bl64_dbg_lib_trace_start
@@ -3201,8 +3202,33 @@ function bl64_fs_run_cp() {
 # Returns:
 #   command exit status
 #######################################
+function bl64_fs_run_rm() {
+  bl64_dbg_lib_show_function "$@"
+  local verbose=''
+
+  bl64_check_parameters_none "$#" || return $?
+  bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_CP_VERBOSE"
+
+  bl64_dbg_lib_trace_start
+  # shellcheck disable=SC2086
+  "$BL64_FS_CMD_RM" $verbose "$@"
+  bl64_dbg_lib_trace_stop
+}
+
+#######################################
+# Command wrapper with verbose, debug and common options
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
 function bl64_fs_run_ls() {
   bl64_dbg_lib_show_function "$@"
+
   bl64_check_parameters_none "$#" || return $?
 
   bl64_dbg_lib_trace_start
@@ -3223,9 +3249,9 @@ function bl64_fs_run_ls() {
 #######################################
 function bl64_fs_run_ln() {
   bl64_dbg_lib_show_function "$@"
-  bl64_check_parameters_none "$#" || return $?
   local verbose=''
 
+  bl64_check_parameters_none "$#" || return $?
   bl64_dbg_lib_command_enabled && verbose="$BL64_FS_SET_LN_VERBOSE"
 
   bl64_dbg_lib_trace_start
@@ -4340,7 +4366,7 @@ function bl64_os_set_lang() {
 #######################################
 # BashLib64 / Module / Functions / OS / Identify OS attributes and provide command aliases
 #
-# Version: 1.19.0
+# Version: 1.20.0
 #######################################
 
 function _bl64_os_match() {
@@ -4431,10 +4457,10 @@ function _bl64_os_get_distro_from_os_release() {
 }
 
 #######################################
-# Check if the current OS matches the target list
+# Compare the current OS version against a list of OS versions
 #
 # Arguments:
-#   $@: each argument is an OS target. Any combintation of the formats: "$BL64_OS_<ALIAS>" "${BL64_OS_<ALIAS>}-V" "${BL64_OS_<ALIAS>}-V.S"
+#   $@: each argument is an OS target. The list is any combintation of the formats: "$BL64_OS_<ALIAS>" "${BL64_OS_<ALIAS>}-V" "${BL64_OS_<ALIAS>}-V.S"
 # Outputs:
 #   STDOUT: None
 #   STDERR: None
@@ -4533,7 +4559,7 @@ function _bl64_os_set_distro() {
 }
 
 #######################################
-# Check if locale resources for language are installed in the OS
+# Determine if locale resources for language are installed in the OS
 #
 # Arguments:
 #   $1: locale name
@@ -4563,6 +4589,32 @@ function bl64_os_lang_is_available() {
   done
 
   return 1
+}
+
+#######################################
+# Check the current OS version is in the supported list
+#
+# * Target use case is script compatibility. Use in the init part to halt execution if OS is not supported
+# * Not recommended for checking individual functions. Instead, use if or case structures to support multiple values based on the OS version
+# * The check is done against the provided list
+# * This is a wrapper to the bl64_os_match so it can be used as a check function
+#
+# Arguments:
+#   $@: list of OS versions to check against. Format: same as bl64_os_match
+# Outputs:
+#   STDOUT: None
+#   STDERR: Error message
+# Returns:
+#   0: check ok
+#   $BL64_LIB_ERROR_APP_INCOMPATIBLE
+#######################################
+function bl64_os_check_version() {
+  bl64_dbg_lib_show_function "$@"
+
+  bl64_os_match "$@" && return 0
+
+  bl64_msg_show_error "${_BL64_OS_TXT_OS_VERSION_NOT_SUPPORTED} (OS: ${BL64_OS_DISTRO} ${BL64_MSG_COSMETIC_PIPE} ${_BL64_OS_TXT_OS_MATRIX}: ${*})"
+  return $BL64_LIB_ERROR_APP_INCOMPATIBLE
 }
 
 #######################################