+++ src/bl64-cnt-env.bash
@@ -1,7 +1,7 @@
 #######################################
 # BashLib64 / Module / Globals / Interact with container engines
 #
-# Version: 1.1.0
+# Version: 1.2.0
 #######################################
 
 # Optional module. Not enabled by default
@@ -9,5 +9,10 @@ export BL64_CNT_MODULE="$BL64_VAR_OFF"
 
 export BL64_CNT_CMD_PODMAN=''
 export BL64_CNT_CMD_DOCKER=''
+
+export BL64_CNT_SET_DOCKER_VERSION=''
+export BL64_CNT_SET_PODMAN_VERSION=''
+
+export BL64_CNT_PATH_DOCKER_SOCKET=''
 
 export _BL64_CNT_TXT_NO_CLI='unable to detect supported container engine'